refactor: tidy some filenames and remaining leftover code
