--- Browser_IDE/Api.js
@@ -1,122 +1,0 @@
-const express = require("express")
-const app = express()
-const bodyP = require("body-parser")
-const compiler = require("compilex")
-var path = require('path');
-const options = { stats: true }
-compiler.init(options)
-app.use(bodyP.json())
-
-app.use("/codemirror-5.65.15", express.static(path.join(__dirname,"node_modules/codemirror")))
-app.use("/jszip", express.static(path.join(__dirname,"node_modules/jszip/dist/")))
-app.use("/babel", express.static(path.join(__dirname,"node_modules/@babel/standalone/")))
-
-app.use(function (req, res, next) {
-    res.header("Access-Control-Allow-Origin", "*");
-    res.header(
-      "Access-Control-Allow-Headers",
-      "Origin, X-Requested-With, Content-Type, Accept"
-    );
-    res.header("Access-Control-Allow-Methods", "GET, POST, OPTIONS, PUT, DELETE");
-    next();
-  });
-
-app.use(express.static(path.join(__dirname,"")));
-
-app.get("/", function (req, res) {
-    compiler.flush(function () {
-        console.log("deleted")
-    })
-
-    res.sendFile("index.html", { root: __dirname })
-})
-app.post("/compile", function (req, res) {
-    var code = req.body.code
-    var input = req.body.input
-    var lang = req.body.lang
-    try {
-
-        if (lang == "Cpp") {
-            if (!input) {
-                var envData = { OS: "windows", cmd: "g++", options: { timeout: 10000 } }; // (uses g++ command to compile )
-                compiler.compileCPP(envData, code, function (data) {
-                    if (data.output) {
-                        res.send(data);
-                    }
-                    else {
-                        res.send({ output: "error" })
-                    }
-                });
-            }
-            else {
-                var envData = { OS: "windows", cmd: "g++", options: { timeout: 10000 } }; // (uses g++ command to compile )
-                compiler.compileCPPWithInput(envData, code, input, function (data) {
-                    if (data.output) {
-                        res.send(data);
-                    }
-                    else {
-                        res.send({ output: "error" })
-                    }
-                });
-            }
-        }
-        else if (lang == "Java") {
-            if (!input) {
-                var envData = { OS: "windows" };
-                compiler.compileJava(envData, code, function (data) {
-                    if (data.output) {
-                        res.send(data);
-                    }
-                    else {
-                        res.send({ output: "error" })
-                    }
-                })
-            }
-            else {
-                //if windows  
-                var envData = { OS: "windows" };
-                //else
-                compiler.compileJavaWithInput(envData, code, input, function (data) {
-                    if (data.output) {
-                        res.send(data);
-                    }
-                    else {
-                        res.send({ output: "error" })
-                    }
-                })
-            }
-        }
-        // Sherap's work
-        else if (lang == "Python") {
-            if (!input) {
-                var envData = { OS: "windows" };
-                
-                compiler.compilePython(envData, code, function (data) {
-                    if (data.output) {
-                        res.send(data);
-                    }
-                    else {
-                        res.send({ output: "error" })
-                    }
-                });
-            }
-            else {
-                var envData = { OS: "windows" };
-                compiler.compilePythonWithInput(envData, code, input, function (data) {
-                    if (data.output) {
-                        res.send(data);
-                    }
-                    else {
-                        res.send({ output: "error" })
-                    }
-                });
-            }
-            // Sherap's work
-        }
-    }
-    catch (e) {
-        console.log("error")
-    }
-})
-
-app.listen(8000)
