+++ Browser_IDE/index.html
@@ -98,7 +98,7 @@
 <script src="IDBStoredProject.js"></script>
 <script src="projectInitializer.js"></script>
 <script src="modal.js"></script>
-<script src="script.js"></script>
+<script src="editorMain.js"></script>
 <script src="treeview.js"></script>
 <script src="fileview.js"></script>
 </html>
